use tag replacement function that is backwards compatible.
this makes use of getTable subsets, so default and variations can be
included in one toml file, to create a Chosen mindset from a series of
mindFrames. Intention is to use overriding for gender, tickle, dignity-
state, etc. the replaceTags function is recursive, so a dignity-
state return string can contain gender tags that will be replaced, later.

Signed-off-by: picobyte <PIc.O.0ob.y.TE@gmail.com>